Added config subcommand to set OpenAI API key and model to use. Improved error handling for config and OpenAI API calls.
